+++ tests/BigQuery.InMemoryEmulator.Tests.Unit/Phase32/InformationSchemaTests.cs
@@ -238,7 +238,7 @@ public class InformationSchemaTests
 
 		Assert.Equal(2, rows.Count);
 		Assert.Equal("id", rows[0].F[2].V?.ToString());
-		Assert.Equal("INTEGER", rows[0].F[3].V?.ToString());
+		Assert.Equal("INT64", rows[0].F[3].V?.ToString());
 		Assert.Equal("name", rows[1].F[2].V?.ToString());
 	}
